pyproject.toml
[project]
name = "lex-os-kernel"
version = "0.1.0"
description = "Middleware de Orquestração Jurídica (Zotero + Fabric + Obsidian)"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.1.0",
    "pyzotero>=1.5.0",
    "pyyaml>=6.0",
    "asyncio"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"